+++ packages/app-admin-ui/client/pages/Item/ItemTitle.js
@@ -21,9 +21,8 @@ const Container = ({ children }) => {
   return <div css={{ paddingLeft: padding, paddingRight: padding }}>{children}</div>;
 };
 
-export const ItemTitle = memo(function ItemTitle({ titleText, adminPath }) {
+export const ItemTitle = memo(function ItemTitle({ titleText }) {
   const { list } = useList();
-  const listHref = `${adminPath}/${list.path}`;
   const { itemHeaderActions } = useUIHooks();
   return (
     <Container>
@@ -34,10 +33,10 @@ export const ItemTitle = memo(function ItemTitle({ titleText, adminPath }) {
           <IconButton
             variant="subtle"
             icon={ChevronLeftIcon}
-            to={listHref}
+            to={list.fullPath}
             css={{ marginLeft: -12 }}
           >
-            Back
+            {list.label}
           </IconButton>
           <Search list={list} />
         </div>