+++ packages/app-admin-ui/client/pages/List/index.js
@@ -2,7 +2,7 @@
 
 import { jsx } from '@emotion/core';
 import { Fragment, useEffect, Suspense } from 'react';
-import { useQuery } from '@apollo/react-hooks';
+import { useHistory, useLocation } from 'react-router-dom';
 import { useList } from '../../providers/List';
 
 import { IconButton } from '@arch-ui/button';
@@ -36,34 +36,19 @@ import { captureSuspensePromises } from '@keystonejs/utils';
 import { useAdminMeta } from '../../providers/AdminMeta';
 
 export function ListLayout(props) {
-  const { items, itemCount, queryErrors, routeProps, query } = props;
+  const { items, itemCount, queryErrors, query } = props;
+
   const { list, openCreateItemModal } = useList();
-  const { urlState } = useListUrlState(list.key);
-  const { filters } = useListFilter(list.key);
-  const [sortBy, handleSortChange] = useListSort(list.key);
+  const {
+    urlState: { currentPage, fields, pageSize, search },
+  } = useListUrlState(list);
 
-  const { adminPath } = useAdminMeta();
-  const { history, location } = routeProps;
-  const { currentPage, fields, pageSize, search } = urlState;
-
+  const { filters } = useListFilter();
+  const [sortBy, handleSortChange] = useListSort();
   const [selectedItems, onSelectChange] = useListSelect(items);
 
-  // Mount with Persisted Search
-  // ------------------------------
-  useEffect(() => {
-    const maybePersistedSearch = list.getPersistedSearch();
-
-    if (location.search) {
-      if (location.search !== maybePersistedSearch) {
-        list.setPersistedSearch(location.search);
-      }
-    } else if (maybePersistedSearch) {
-      history.replace({
-        ...location,
-        search: maybePersistedSearch,
-      });
-    }
-  }, []);
+  const { adminPath } = useAdminMeta();
+  const history = useHistory();
 
   // Misc.
   // ------------------------------
@@ -81,7 +66,7 @@ export function ListLayout(props) {
   const onCreate = ({ data }) => {
     const id = data[list.gqlNames.createMutationName].id;
     query.refetch().then(() => {
-      history.push(`${adminPath}/${list.path}/${id}`);
+      history.push(`${list.fullPath}/${id}`);
     });
   };
 
@@ -163,14 +148,13 @@ export function ListLayout(props) {
                   {sortBy ? (
                     <Fragment>
                       <span css={{ paddingLeft: '0.5ex' }}>sorted by</span>
-                      <SortPopout listKey={list.key} />
+                      <SortPopout />
                     </Fragment>
                   ) : (
                     ''
                   )}
                   <span css={{ paddingLeft: '0.5ex' }}>with</span>
                   <ColumnPopout
-                    listKey={list.key}
                     target={handlers => (
                       <Button
                         variant="subtle"
@@ -194,7 +178,7 @@ export function ListLayout(props) {
                             .filter(field => field.path !== '_label_')
                             .map(field => () => field.initCellView())
                         );
-                        return <Pagination listKey={list.key} isLoading={query.loading} />;
+                        return <Pagination isLoading={query.loading} />;
                       }}
                     </Render>
                   </Suspense>
@@ -253,18 +237,15 @@ export function ListLayout(props) {
   );
 }
 
-export function List(props) {
-  const { list, query, routeProps } = props;
+const ListPage = props => {
+  const {
+    list,
+    listData: { items, itemCount, queryErrors },
+    query,
+  } = useList();
 
-  // get item data
-  const items = query.data && query.data[list.gqlNames.listQueryName];
-  const queryErrors = query.data && query.data.error;
-  let itemCount;
-  if (query.data && query.data[list.gqlNames.listQueryMetaName]) {
-    itemCount = query.data[list.gqlNames.listQueryMetaName].count;
-  }
-
-  const { history, location } = routeProps;
+  const history = useHistory();
+  const location = useLocation();
 
   // Mount with Persisted Search
   // ------------------------------
@@ -287,16 +268,12 @@ export function List(props) {
   // ------------------------------
   // Only show error page if there is no data
   // (ie; there could be partial data + partial errors)
-  if (
-    query.error &&
-    (!query.data ||
-      !query.data[list.gqlNames.listQueryName] ||
-      !Object.keys(query.data[list.gqlNames.listQueryName]).length)
-  ) {
+  // Note this is the error returned from Apollo, *not* any that are part of the GQL result.
+  if (query.error && (!query.data || !items || !Object.keys(items).length)) {
     let message = query.error.message;
 
-    // If there was an error returned by GraphQL, use that message
-    // instead
+    // If there was an error returned by GraphQL, use that message instead
+    // FIXME: convert this to an optional chaining operator at some point
     if (
       query.error.networkError &&
       query.error.networkError.result &&
@@ -333,19 +310,6 @@ export function List(props) {
       />
     </Fragment>
   );
-}
+};
 
-export default function ListData(props) {
-  const { list } = props;
-  const { urlState } = useListUrlState(list.key);
-
-  const { currentPage, fields, filters, pageSize, search, sortBy } = urlState;
-  const orderBy = sortBy ? `${sortBy.field.path}_${sortBy.direction}` : null;
-  const first = pageSize;
-  const skip = (currentPage - 1) * pageSize;
-
-  const query = list.getQuery({ fields, filters, search, orderBy, skip, first });
-  const res = useQuery(query, { fetchPolicy: 'cache-and-network', errorPolicy: 'all' });
-
-  return <List query={res} {...props} />;
-}
+export default ListPage;