+++ packages/fields/src/types/Decimal/views/Field.js
@@ -37,7 +37,7 @@ const DecimalField = ({ onChange, autoFocus, field, value, errors }) => {
   return (
     <FieldContainer>
       <FieldLabel htmlFor={htmlID} field={field} errors={errors} />
-      <FieldDescription text={field.config.adminDoc} />
+      <FieldDescription text={field.adminDoc} />
       <FieldInput>
         {symbol && <Currency>{symbol}</Currency>}
         <Input