+++ packages/fields/src/types/Relationship/views/Controller.js
@@ -2,13 +2,11 @@ import FieldController from '../../../Controller';
 
 export default class RelationshipController extends FieldController {
   constructor(config, ...args) {
-    const defaultValue = 'defaultValue' in config ? config.defaultValue : config.many ? [] : null;
+    const { defaultValue = config.many ? [] : null } = config;
     super({ ...config, defaultValue }, ...args);
   }
   getRefList() {
-    const { getListByKey } = this.adminMeta;
-    const { ref } = this.config;
-    return getListByKey(ref);
+    return this.getListByKey(this.config.ref);
   }
   getQueryFragment = (path = this.path) => {
     return `