+++ packages/keystone/tests/List.test.js
@@ -250,6 +250,7 @@ describe('new List()', () => {
       listQueryName: 'allTests',
       listQueryMetaName: '_allTestsMeta',
       listMetaName: '_TestsMeta',
+      listSortName: 'SortTestsBy',
       deleteMutationName: 'deleteTest',
       deleteManyMutationName: 'deleteTests',
       updateMutationName: 'updateTest',
@@ -438,6 +439,7 @@ describe('getAdminMeta()', () => {
       itemQueryName: 'Test',
       listQueryName: 'allTests',
       listQueryMetaName: '_allTestsMeta',
+      listSortName: 'SortTestsBy',
       listMetaName: '_TestsMeta',
       deleteMutationName: 'deleteTest',
       deleteManyMutationName: 'deleteTests',
@@ -473,42 +475,6 @@ describe('getAdminMeta()', () => {
     expect(adminMeta.fields[2].path).toEqual('email');
     expect(adminMeta.fields[3].path).toEqual('other');
     expect(adminMeta.fields[4].path).toEqual('writeOnce');
-  });
-
-  test('getAdminMeta() - views', () => {
-    const list = setup();
-    const schemaName = 'public';
-    const adminMeta = list.getAdminMeta({ schemaName });
-
-    expect(adminMeta.views).toEqual({
-      id: {}, // Mocked
-      name: {
-        Controller: resolveViewPath('Text/views/Controller'),
-        Field: resolveViewPath('Text/views/Field'),
-        Filter: resolveViewPath('Text/views/Filter'),
-      },
-      email: {
-        Controller: resolveViewPath('Text/views/Controller'),
-        Field: resolveViewPath('Text/views/Field'),
-        Filter: resolveViewPath('Text/views/Filter'),
-      },
-      other: {
-        Controller: resolveViewPath('Relationship/views/Controller'),
-        Field: resolveViewPath('Relationship/views/Field'),
-        Filter: resolveViewPath('Relationship/views/Filter'),
-        Cell: resolveViewPath('Relationship/views/Cell'),
-      },
-      hidden: {
-        Controller: resolveViewPath('Text/views/Controller'),
-        Field: resolveViewPath('Text/views/Field'),
-        Filter: resolveViewPath('Text/views/Filter'),
-      },
-      writeOnce: {
-        Controller: resolveViewPath('Text/views/Controller'),
-        Field: resolveViewPath('Text/views/Field'),
-        Filter: resolveViewPath('Text/views/Filter'),
-      },
-    });
   });
 });
 
@@ -613,7 +579,16 @@ describe(`getGqlTypes() `, () => {
   const createManyInput = `input TestsCreateInput {
         data: TestCreateInput
       }`;
-
+  const sortTestsBy = `enum SortTestsBy {
+        name_ASC
+        name_DESC
+        email_ASC
+        email_DESC
+        other_ASC
+        other_DESC
+        writeOnce_ASC
+        writeOnce_DESC
+      }`;
   const schemaName = 'public';
   test('access: true', () => {
     expect(
@@ -625,6 +600,7 @@ describe(`getGqlTypes() `, () => {
         type,
         whereInput,
         whereUniqueInput,
+        sortTestsBy,
         updateInput,
         updateManyInput,
         createInput,
@@ -644,7 +620,7 @@ describe(`getGqlTypes() `, () => {
       setup({ access: { read: true, create: false, update: false, delete: false } })
         .getGqlTypes({ schemaName })
         .map(s => print(gql(s)))
-    ).toEqual([type, whereInput, whereUniqueInput].map(s => print(gql(s))));
+    ).toEqual([type, whereInput, whereUniqueInput, sortTestsBy].map(s => print(gql(s))));
   });
   test('create: true', () => {
     expect(
@@ -652,7 +628,9 @@ describe(`getGqlTypes() `, () => {
         .getGqlTypes({ schemaName })
         .map(s => print(gql(s)))
     ).toEqual(
-      [type, whereInput, whereUniqueInput, createInput, createManyInput].map(s => print(gql(s)))
+      [type, whereInput, whereUniqueInput, sortTestsBy, createInput, createManyInput].map(s =>
+        print(gql(s))
+      )
     );
   });
   test('update: true', () => {
@@ -661,7 +639,9 @@ describe(`getGqlTypes() `, () => {
         .getGqlTypes({ schemaName })
         .map(s => print(gql(s)))
     ).toEqual(
-      [type, whereInput, whereUniqueInput, updateInput, updateManyInput].map(s => print(gql(s)))
+      [type, whereInput, whereUniqueInput, sortTestsBy, updateInput, updateManyInput].map(s =>
+        print(gql(s))
+      )
     );
   });
   test('delete: true', () => {
@@ -669,7 +649,7 @@ describe(`getGqlTypes() `, () => {
       setup({ access: { read: false, create: false, update: false, delete: true } })
         .getGqlTypes({ schemaName })
         .map(s => print(gql(s)))
-    ).toEqual([type, whereInput, whereUniqueInput].map(s => print(gql(s))));
+    ).toEqual([type, whereInput, whereUniqueInput, sortTestsBy].map(s => print(gql(s))));
   });
 });
 
@@ -678,6 +658,7 @@ test('getGraphqlFilterFragment', () => {
   expect(list.getGraphqlFilterFragment()).toEqual([
     'where: TestWhereInput',
     'search: String',
+    'sortBy: [SortTestsBy!]',
     'orderBy: String',
     'first: Int',
     'skip: Int',
@@ -697,6 +678,7 @@ describe(`getGqlQueries()`, () => {
           allTests(
           where: TestWhereInput
           search: String
+          sortBy: [SortTestsBy!]
           orderBy: String
           first: Int
           skip: Int
@@ -709,6 +691,7 @@ describe(`getGqlQueries()`, () => {
           _allTestsMeta(
           where: TestWhereInput
           search: String
+          sortBy: [SortTestsBy!]
           orderBy: String
           first: Int
           skip: Int
@@ -877,7 +860,7 @@ test('checkFieldAccess', () => {
       'read',
       [{ existingItem: { makeFalse: true }, data: { name: 'a', email: 'a@example.com' } }],
       context,
-      { gqlName: 'testing', extraData: { extra: 1 } }
+      { gqlName: 'testing', extra: 1 }
     );
   } catch (error) {
     thrownError = error;